Remove ContributionAppealController and associated classes
This commit removes the _ContributionAppealController_ and associated classes.
This controller was previously called by the Crown Court Contribution service
to calculate appeal contribution amounts, but this call was recently removed
(see [CCC PR 159](https://github.com/ministryofjustice/laa-crown-court-contribution/pull/159))
in favour of encapsulating the calculation in an enum to remove this cross-service
call.