Add feature flag memberships to user summary information
This commit adds feature flag memberships to the user summary information,
both for user-specific memberships and for those which apply to all users.

This update is made to allow for changes in services which use feature flags
to control logic and specifically, to remove the need to pass feature flag
information between services (such as between MAAT and the Orchestration
Service). The feature flag information is added specifically to the
`UserSummaryDTO` as this is the most appropriate place to hold this information,
similar to that held already for role data items.