Filter feature toggles to prioritise user-specific toggles
This commit adds filtering of feature toggles that are added to the `UserSummaryDTO`
to ensure that there are no duplicates between user-specific feature toggles and
those toggles which are applied to all users (where username is `null`).

In situations where this duplication exists, priority is given to the user-specific
toggle where the feature and action are equal to the more generic version; for feature
toggles available to all users where there is no exact match to feature and action for
a user-specific toggle, these toggles are also made available.

The decision to encapsulate this logic in code rather than in a database query means
that it is more testable and easier to understand than what would otherwise have been
a multi-statement query to try and achieve the same output. It can also be argued that
application-specific filtering should be done in the application rather than in the
database to make better use of these specific resource types.