Upgrade Spring Boot and other dependencies
This PR upgrades Spring Boot from version `3.1.5` to `3.4.1`, as well as
upgrading many other dependencies as a result to ensure that everything
remains compatible. This upgrade has brought about some required changes
to the codebase, both as a result of behavioural changes with the h2database
package and also from changes to Spring Boot and HIbernate around optimistic
locking (where an exception is now thrown by a downstream package when it
wasn't before), giving visibility into what was an existing issue.

One change to production code to call out is that to FinancialAssessmentHistoryImpl.
Previously, a new FinancialAssessmentEntity was being created as part of the process
creating a FinancialHistoryAssessmentEntity to save to the database, however because
the FinancialAssessmentEntity had not been persisted itself to the database, saving
the FinancialHistoryAssessmentEntity resulted in a
org.hibernate.TransientPropertyValueException exception because of the unsaved
FinancialAssessmentEntity. Upon inspection, the code takes the
FinancialAssessmentEntity, maps it to a FinancialAssessmentDTO (to save into the
FinancialAssessmentsHistoryDTO) after which the FinancialAssessmentDTO is then mapped
back to a (new) FinancialAssessmentEntity; instead, the existing FinancialAssessmentEntity
is now being assigned to the FinancialHistoryAssessmentEntity once it has been created,
overwriting the entity has been created during mapping.