Fix bug when patching entity with a LocalDateTime
This commit fixes an existing and previously uncovered bug with
multiple patch and update endpoints in the Court Data API, which
occurs when attempting to set a _LocalDateTime_ object passed as
part of the fields to patch against the respective field on the
entity object attempting to be patched. It does this by
specifically checking the type of the field attempting to be
set on the entity object and parsing the _LocalDateTime_ from the
String if this is the type of the incoming field to update.

This approach is similar to an existing approach in the Prison API
which tackles a similar problem: https://github.com/ministryofjustice/prison-api/blob/0003f5487e1225dd30157dce1d17da8101693a17/src/main/java/uk/gov/justice/hmpps/prison/repository/mapping/FieldMapper.java#L51.

This bug was likely previously not discovered as there were no
tests in place attempting to set a _LocalDateTime_ instance on
an entity, only Strings, longs and ints. It should also be noted
that this bug would only present itself when passing the fields
to update into the Court Data API via the controller; internal
calls to patch or update an entity with a _LocalDateTime_ field
would work fine as the update field in this scenario would remain
a _LocalDateTime_ instance rather than a String.

This commit also refactors existing conditional logic to make it
simpler and remove unnecessary overhead.