Add endpoint to reset rep orders for billing
This commit adds a new endpoint to reset rep orders for billing,
setting the `SEND_TO_CCLF` flag on relevant rep orders in the
database to `null` and updating the date and user modified values
alongside.

This behaviour is migrated from the existing `reporder_clear_flag`
stored procedure in the MAAT database which was previously called
as part of the regular data extraction Hub job.