Add new RepOrderBillingEntity
This commit adds a new `RepOrderBillingEntity` alongside
a new `RepOrderBillingRepository`, for the process of
extracting data for CCLF and updating extracted records.

The `RepOrderEntity` was previously used for this
purpose, but was failing at runtime due to the fact that
not all data expected by the entity was being pulled back
by the query. By introducing a new entity type, the query
can be kept as-is to return only the necessary data for
the data extract.